Implementation for Discrete events (selection, move binary to descrete, some event query rewrite)

+++ subprojects/probability/src/main/java/tools/refinery/probability/terms/EventHandle.java
@@ -4,18 +4,20 @@ import tools.refinery.logic.term.ConstantTerm;
 import tools.refinery.logic.term.Term;
 
 public class EventHandle {
-	private int handle;
-	public EventHandle(int id){
+	private double handle;
+	public EventHandle(double id){
 		this.handle = id;
 	}
-	public int getValue(){
+	public double getValue(){
 		return handle;
 	}
-	public static Term<EventHandle> of(int id){
+
+	public static Term<EventHandle> of(double id){
 		return new ConstantTerm<EventHandle>(EventHandle.class, new EventHandle(id));
 	}
 	@Override
 	public String toString(){
-		return "%d".formatted(handle);
+		return "%f".formatted(handle);
 	}
+
 }
